Fix for links in site_pages_url
